Add dedicated solo and duo rank matching pages
